simple-agent/requirements.txt
azure-ai-projects==2.0.1
azure-core>=1.37.0
azure-identity==1.20.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.1
idna==3.10
isodate==0.7.2
msal==1.31.1
msal-extensions==1.2.0
portalocker==2.10.1
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.0.1
pywin32==308
requests==2.32.3
six==1.17.0
typing_extensions==4.12.2
urllib3==2.3.0
